// src/email-marketing/enums/segment-rule-operator.enum.ts
export enum SegmentRuleOperator {
    EQUALS = 'equals',
    NOT_EQUALS = 'not_equals',
    CONTAINS = 'contains',
    NOT_CONTAINS = 'not_contains',
    STARTS_WITH = 'starts_with',
    ENDS_WITH = 'ends_with',
    GREATER_THAN = 'greater_than',
    LESS_THAN = 'less_than',
    GREATER_OR_EQUAL = 'greater_or_equal',
    LESS_OR_EQUAL = 'less_or_equal',
    IS_SET = 'is_set',
    IS_NOT_SET = 'is_not_set',
    IN_LIST = 'in_list',
    NOT_IN_LIST = 'not_in_list',
    BEFORE = 'before',
    AFTER = 'after',
    BETWEEN = 'between',
}