// src/email-marketing/enums/segment-rule-field.enum.ts
export enum SegmentRuleField {
    EMAIL = 'email',
    FIRST_NAME = 'first_name',
    LAST_NAME = 'last_name',
    CREATED_AT = 'created_at',
    LAST_LOGIN = 'last_login',
    COURSE_COUNT = 'course_count',
    TOTAL_SPENT = 'total_spent',
    TAG = 'tag',
    COUNTRY = 'country',
    SUBSCRIPTION_STATUS = 'subscription_status',
}